Move imagesRouteToApiUrl into api.ts
