+++ common/services/catalogue/ts_api.ts
@@ -57,22 +57,3 @@ export type CatalogueWorksApiProps = {
   _queryType?: string;
   aggregations?: string[];
 };
-
-export function imagesRouteToApiUrl(
-  imagesRouteProps: ImagesProps,
-  overrides: Partial<CatalogueImagesApiProps>
-): CatalogueImagesApiProps {
-  return {
-    query: imagesRouteProps.query,
-    page: imagesRouteProps.page,
-    color: imagesRouteProps.color,
-    'locations.license': imagesRouteProps['locations.license'],
-    'source.genres.label':
-      imagesRouteProps['source.genres.label'].map(quoteVal),
-    'source.subjects.label':
-      imagesRouteProps['source.subjects.label'].map(quoteVal),
-    'source.contributors.agent.label':
-      imagesRouteProps['source.contributors.agent.label'].map(quoteVal),
-    ...overrides,
-  };
-}
